fixing k_dmgfrags, now u get frags when do damage to pented enemy too, hope it work, doesn't tested.

+++ include/progs.h
@@ -112,7 +112,7 @@ typedef struct player_stats_s {
 	float    dmg_t; // damage taken
 	float    dmg_g; // damage given
 	float    dmg_team;  // damage to team
-	int      dmg_frags; // frags awarded from damage (CA)
+	float    dmg_frags; // frags awarded from damage (CA)
 
 	int		ot_a;	 // overtime armor value
 //	float	ot_at;   // overtime armor type
